Handle clicking on a label inside `UnifiedResources` (#35255)
* Extract `makeAdvancedSearchQueryForLabel` function

* Encapsulate all logic related to clicking on a label inside the shared `UnifiedResources`

* Run linter

(cherry picked from commit 81090e2ee824fde745448323ad3d35a043811939)
