+++ web/packages/shared/components/UnifiedResources/UnifiedResources.story.tsx
@@ -128,7 +128,6 @@ const story = ({
         pinning={pinning}
         unifiedResourcePreferences={userPrefs}
         updateUnifiedResourcesPreferences={setUserPrefs}
-        onLabelClick={() => undefined}
         NoResources={undefined}
         fetchResources={fetch}
         resourcesFetchAttempt={attempt}
