+++ web/packages/shared/components/UnifiedResources/UnifiedResources.tsx
@@ -30,7 +30,7 @@ import { Danger } from 'design/Alert';
 
 import './unifiedStyles.css';
 
-import { ResourcesResponse, ResourceLabel } from 'teleport/services/agents';
+import { ResourcesResponse } from 'teleport/services/agents';
 
 import {
   DefaultTab,
@@ -49,6 +49,7 @@ import {
   useInfiniteScroll,
 } from 'shared/hooks/useInfiniteScroll';
 import { Attempt } from 'shared/hooks/useAttemptNext';
+import { makeAdvancedSearchQueryForLabel } from 'shared/utils/advancedSearchLabelQuery';
 
 import {
   SharedUnifiedResource,
@@ -132,7 +133,6 @@ interface UnifiedResourcesProps {
   pinning: UnifiedResourcesPinning;
   availableKinds: FilterKind[];
   setParams(params: UnifiedResourcesQueryParams): void;
-  onLabelClick(label: ResourceLabel): void;
   /** A list of actions that can be performed on the selected items. */
   bulkActions?: BulkAction[];
   unifiedResourcePreferences: UnifiedResourcePreferences;
@@ -148,7 +148,6 @@ export function UnifiedResources(props: UnifiedResourcesProps) {
     resourcesFetchAttempt,
     resources,
     fetchResources,
-    onLabelClick,
     availableKinds,
     pinning,
     unifiedResourcePreferences,
@@ -440,7 +439,13 @@ export function UnifiedResources(props: UnifiedResourcesProps) {
         <PinningNotSupported />
       ) : (
         <ViewComponent
-          onLabelClick={onLabelClick}
+          onLabelClick={label =>
+            setParams({
+              ...params,
+              search: '',
+              query: makeAdvancedSearchQueryForLabel(label, params),
+            })
+          }
           pinnedResources={pinnedResources}
           selectedResources={selectedResources}
           onSelectResource={handleSelectResource}
